refactor(7161): isolate settlement authority construction
Keep escrow state transitions bounded by moving canonical bridge authority projection into a dedicated module.

Constraint: Preserve the locked authority response byte shape

Confidence: high

Scope-risk: narrow

Directive: Derive receipt-chain commitments through task projection rather than duplicating the algorithm

Tested: focused authority contracts and clippy with warnings denied
